+++ src/hud/sidebar/skills.ts
@@ -46,10 +46,6 @@ const SHARED_ACTIONS = {
         trained: true,
         uuid: "Compendium.pf2e.actionspf2e.Item.Q5iIYCFdqJFM31GW",
     },
-    earnIncome: {
-        trained: true,
-        uuid: "Compendium.pf2e.actionspf2e.Item.QyzlsLrqM0EEwd7j",
-    },
     subsist: {
         uuid: "Compendium.pf2e.actionspf2e.Item.49y9Ec4bDii8pcD3",
     },
@@ -192,7 +188,6 @@ const SKILLS: RawSkill[] = [
                 trained: true,
                 uuid: "Compendium.pf2e.actionspf2e.Item.rmwa3OyhTZ2i2AHl",
             },
-            "earnIncome",
             {
                 actionId: "identify-alchemy",
                 trained: true,
@@ -367,7 +362,6 @@ const SKILLS: RawSkill[] = [
                 ],
                 uuid: "Compendium.pf2e.actionspf2e.Item.EEDElIyin4z60PXx",
             },
-            "earnIncome",
         ],
     },
     {
@@ -443,6 +437,10 @@ const SKILLS: RawSkill[] = [
     {
         slug: "stealth",
         actions: [
+            {
+                actionId: "avoid-notice",
+                uuid: "Compendium.pf2e.actionspf2e.Item.IE2nThCmoyhQA0Jn",
+            },
             {
                 actionId: "conceal-an-object",
                 cost: 1,