+++ src/hud/tracker.ts
@@ -2,7 +2,6 @@ import {
     R,
     addListener,
     addListenerAll,
-    canObserveActor,
     createHook,
     elementDataset,
     getDispositionColor,
@@ -18,7 +17,7 @@ import {
 } from "foundry-pf2e";
 import Sortable, { SortableEvent } from "sortablejs";
 import { BaseRenderOptions, BaseSettings, PF2eHudBase } from "./base/base";
-import { HealthData, getHealth } from "./shared/base";
+import { HealthData, getHealth, userCanObserveActor } from "./shared/base";
 
 // Hooks.on("getSceneControlButtons", (controls) => {
 //     controls[0].tools.push({
@@ -34,7 +33,7 @@ import { HealthData, getHealth } from "./shared/base";
 //     });
 // });
 
-class PF2eHudTracker extends PF2eHudBase<TrackerSettings> {
+class PF2eHudTracker extends PF2eHudBase<TrackerSettings, any, TrackerRenderOptions> {
     #hoverTokenHook = createHook("hoverToken", this.#onHoverToken.bind(this));
     #targetTokenHook = createHook("targetToken", this.#refreshTargetDisplay.bind(this));
     #renderEffectsHook = createHook("renderEffectsPanel", this.#onRenderEffectsPanel.bind(this));
@@ -136,8 +135,11 @@ class PF2eHudTracker extends PF2eHudBase<TrackerSettings> {
         this.#combatTrackerHook.toggle(enabled);
         this.#renderEffectsHook.toggle(enabled);
 
-        if (enabled && this.combat) this.render(true);
-        else if (!enabled && this.rendered) this.close();
+        if (enabled && this.combat) {
+            this.render(true);
+        } else if (!enabled && this.rendered) {
+            this.close();
+        }
 
         // toggleControlTool("pf2e-hud-tracker", enabled);
 
@@ -182,10 +184,17 @@ class PF2eHudTracker extends PF2eHudBase<TrackerSettings> {
             const playersCanSeeName = !tokenSetsNameVisibility || combatant.playersCanSeeName;
             const dispositionColor = getDispositionColor(actor);
 
-            const texture = {
+            const texture: TrackerTexture = {
                 ...((useTextureScaling && combatant.token?.texture) || { scaleX: 1, scaleY: 1 }),
                 img: await this.tracker._getCombatantThumbnail(combatant),
             };
+
+            if (texture.scaleX >= 1.2 || texture.scaleY >= 1.2) {
+                const scale = texture.scaleX > texture.scaleY ? texture.scaleX : texture.scaleY;
+                const ringPercent = 100 - Math.floor(((scale - 0.7) / scale) * 100);
+                const limitPercent = 100 - Math.floor(((scale - 0.8) / scale) * 100);
+                texture.mask = `radial-gradient(circle at center, black ${ringPercent}%, rgba(0, 0, 0, 0.2) ${limitPercent}%)`;
+            }
 
             const toggleName = (() => {
                 if (!isGM || !tokenSetsNameVisibility || actor?.alliance === "party") return;
@@ -218,7 +227,7 @@ class PF2eHudTracker extends PF2eHudBase<TrackerSettings> {
                 canPing: canPing && combatant.sceneId === canvas.scene?.id,
                 css: css.join(" "),
                 isOwner,
-                health: actor && canObserveActor(actor) ? getHealth(actor) : undefined,
+                health: actor && userCanObserveActor(actor) ? getHealth(actor) : undefined,
                 color: dispositionColor.rgb.map((x) => x * 255).join(", "),
             };
 
@@ -649,6 +658,13 @@ class PF2eHudTracker extends PF2eHudBase<TrackerSettings> {
 
 type EventAction = "toggle-expand" | "delay-turn";
 
+type TrackerTexture = {
+    scaleX: number;
+    scaleY: number;
+    img: string;
+    mask?: string;
+};
+
 type TrackerTurn = {
     id: string;
     index: number;
@@ -662,7 +678,7 @@ type TrackerTurn = {
     hasRolled: boolean;
     health: HealthData | undefined;
     css: string;
-    texture: { scaleX: number; scaleY: number; img: string };
+    texture: TrackerTexture;
     toggleName: Maybe<{
         tooltip: string;
         active: boolean;